\name{IlluminaDataTestFiles-package}
\alias{IlluminaDataTestFiles-package}
\alias{IlluminaDataTestFiles}
\docType{package}
\title{
  Illumina microarray files (IDAT) for testing
}
\description{
  Example data for Illumina microarray output files, for testing purposes
}
\details{
 The package contain public IDAT files from different sources,
 representing both encrypted and unencrypted IDAT files.

}
\author{
Kasper D. Hansen, Mike L. Smith
}
\keyword{ package }
\seealso{
  The illuminaio package for parsing utilities.
}
